% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{glc_resources}
\alias{glc_resources}
\title{Inventory data-package resources}
\usage{
glc_resources(x)
}
\arguments{
\item{x}{A package opened with \code{\link[=glc_open]{glc_open()}}.}
}
\value{
A tibble with one row per declared resource path.
}
\description{
Inventory data-package resources
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
iztech <- glc_open("tscnlab/melidos-iztech-glc-dataset")
glc_resources(iztech)
\dontshow{\}) # examplesIf}
}
