% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{glc_schema_versions}
\alias{glc_schema_versions}
\title{Report supported GLC schema versions}
\usage{
glc_schema_versions()
}
\value{
A tibble describing support status for each schema version.
}
\description{
Report supported GLC schema versions
}
\examples{
glc_schema_versions()
}
