% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glcdp-package.R
\docType{package}
\name{glcdp-package}
\alias{glcdp}
\alias{glcdp-package}
\title{glcdp: Global Light Commons data-package helpers}
\description{
\code{glcdp} discovers, inspects, downloads, and imports Global Light Commons
data packages. Remote packages are resolved at immutable commits, and data
are imported from metadata-described file groups.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tscnlab.github.io/glc-dp-r/}
  \item \url{https://github.com/tscnlab/glc-dp-r}
  \item Report bugs at \url{https://github.com/tscnlab/glc-dp-r/issues}
}

}
\author{
\strong{Maintainer}: Johannes Zauner \email{johannes.zauner@tum.de} (\href{https://orcid.org/0000-0003-2171-4566}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Johannes Zauner \email{johannes.zauner@tum.de} (\href{https://orcid.org/0000-0003-2171-4566}{ORCID}) [copyright holder]
  \item Salma M. Thalji \email{salma.thalji@tum.de} (\href{https://orcid.org/0009-0002-7320-9289}{ORCID}) [copyright holder]
  \item Manuel Spitschan \email{manuel.spitschan@tum.de} (\href{https://orcid.org/0000-0002-8572-9268}{ORCID}) [copyright holder]
}

}
\keyword{internal}
