% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{glc_datasets}
\alias{glc_datasets}
\title{Inventory datasets}
\usage{
glc_datasets(x, dataset_id = NULL)
}
\arguments{
\item{x}{A package opened with \code{\link[=glc_open]{glc_open()}}.}

\item{dataset_id}{Optional dataset id or ids.}
}
\value{
A tibble with one row per dataset.
}
\description{
Inventory datasets
}
